refactor(gateway): move resource URI domain helpers to core
